jpa 설정 및 userEntity 생성
